requirements.txt
# Core LLM
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain>=0.2.0

# Tools
tavily-python>=0.3.0
langchain-chroma>=0.1.0

# Document processing
pypdf>=4.0.0
docx2txt>=0.8

# Database
langchain-community[sql]>=0.2.0

# Utilities
python-dotenv>=1.0.0
sentence-transformers>=2.7.0

# Development (optional)
pytest>=8.0.0
black>=24.0.0